Move the ADR-002 consume guard into its own module
`test_rust_streams.py` carried two unrelated concerns: the behaviour of the
Rust pump and consume entry points, and an AST scan asserting that production
code does not yet route through `rust_consume_stream`. The scan never touches
the compiled extension, so it does not belong in an extension-gated module,
and at 394 lines the file had no room left under the 400-line cap for the
consume-side I/O-error coverage issue #276 asks for.

Lift the scanner, its self-test, and the two status guards into
`test_rust_consume_integration_guard.py`, and repoint the ADR and roadmap
references at the new home. The guards keep running under the ordinary test
glob; they are deliberately left outwith `EXTENSION_TEST_TARGETS` because they
have no dependency on the extension.
